Calibrate

General calibration function for malariasimulation.

Usage

calibrate(
  target,
  summary_function,
  parameters,
  tolerance,
  weights = rep(1, length(target)),
  elimination_check = TRUE,
  maxiter = 20,
  low = 0.001,
  high = 2000
)

Arguments

target

Values of target variable to calibrate to.

summary_function

A function that take the raw model output as an argument and produces a vector of the target variable.

parameters

Other malariasimulation model parameters

tolerance

The routine will complete when the absolute sum of the weighted difference between the target variable and the target values falls below this value. Tolerance is specified in the units of the target variable (e.g. if my target variable is prevalence, and my tolerance is 0.05, then the absolute sum of the weighted difference between the target variable and the target values must be <5% for the routine to succeed).

weights

A numerical vector of weights the same length as target giving the weights to use for elements of target.

elimination_check

Check transmission is maintained for all target points with ongoing transmission before exiting early.

maxiter

Maximum iterations. If stopping criteria have not been met by maxiter, the current best estimate will be returned

low

Lower boundof EIRs

high

Upper bound of EIRs

Value

Uniroot output